fixing outdated version of Leo initial state randomization

+++ addons/rbdl/include/grl/environments/leo/rbdl_leo_task.h
@@ -97,7 +97,7 @@ class LeoSquattingTask : public Task
 
   public:
     LeoSquattingTask() : target_env_(NULL), timeout_(0), weight_nmpc_(0.0001), weight_nmpc_aux_(1.0), weight_nmpc_qd_(1.0), weight_shaping_(0.0),
-      use_mef_(0), task_reward_(0.0), subtask_reward_(0.0), power_(2.0), randomize_(0), dof_(4), continue_after_fall_(0), setpoint_reward_(1), gamma_(0.97),
+      use_mef_(0), task_reward_(0.0), subtask_reward_(0.0), power_(2.0), randomize_(0.), dof_(4), continue_after_fall_(0), setpoint_reward_(1), gamma_(0.97),
       fixed_arm_(false), lower_height_(0.28), upper_height_(0.35), friction_compensation_(false) { }
 
     // From Configurable
@@ -124,7 +124,7 @@ class LeoSquattingTask : public Task
     mutable double subtask_reward_;
     mutable std::vector<double> subtasks_rewards_;
     double power_;
-    int randomize_;
+    double randomize_;
     int dof_;
     Vector target_obs_min_, target_obs_max_;
     int continue_after_fall_;
